// mollys-movies-ui/stubs/api/model/live-transmission-status.ts
// tslint:disable
/**
 * Public Molly's Movies API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface LiveTransmissionStatus
 */
export interface LiveTransmissionStatus {
    /**
     * 
     * @type {string}
     * @memberof LiveTransmissionStatus
     */
    name?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof LiveTransmissionStatus
     */
    complete?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof LiveTransmissionStatus
     */
    stalled?: boolean | null;
    /**
     * 
     * @type {number}
     * @memberof LiveTransmissionStatus
     */
    eta?: number | null;
    /**
     * 
     * @type {number}
     * @memberof LiveTransmissionStatus
     */
    percentComplete?: number | null;
}