// mollys-movies-ui/stubs/api/model/local-movie-source.ts
// tslint:disable
/**
 * Public Molly's Movies API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface LocalMovieSource
 */
export interface LocalMovieSource {
    /**
     * 
     * @type {string}
     * @memberof LocalMovieSource
     */
    source?: string | null;
    /**
     * 
     * @type {string}
     * @memberof LocalMovieSource
     */
    dateCreated?: string;
    /**
     * 
     * @type {string}
     * @memberof LocalMovieSource
     */
    dateScraped?: string;
}