// mollys-movies-ui/stubs/api/model/scrape.ts
// tslint:disable
/**
 * Public Molly's Movies API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { ScrapeSource } from './scrape-source';

/**
 * 
 * @export
 * @interface Scrape
 */
export interface Scrape {
    /**
     * 
     * @type {string}
     * @memberof Scrape
     */
    id?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Scrape
     */
    startDate?: string;
    /**
     * 
     * @type {string}
     * @memberof Scrape
     */
    endDate?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof Scrape
     */
    success?: boolean | null;
    /**
     * 
     * @type {number}
     * @memberof Scrape
     */
    localMovieCount?: number;
    /**
     * 
     * @type {number}
     * @memberof Scrape
     */
    movieCount?: number;
    /**
     * 
     * @type {number}
     * @memberof Scrape
     */
    torrentCount?: number;
    /**
     * 
     * @type {Array<ScrapeSource>}
     * @memberof Scrape
     */
    sources?: Array<ScrapeSource> | null;
}