// mollys-movies-ui/stubs/api/model/scrape-source.ts
// tslint:disable
/**
 * Public Molly's Movies API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { ScraperType } from './scraper-type';

/**
 * 
 * @export
 * @interface ScrapeSource
 */
export interface ScrapeSource {
    /**
     * 
     * @type {string}
     * @memberof ScrapeSource
     */
    source?: string | null;
    /**
     * 
     * @type {ScraperType}
     * @memberof ScrapeSource
     */
    type?: ScraperType;
    /**
     * 
     * @type {boolean}
     * @memberof ScrapeSource
     */
    success?: boolean | null;
    /**
     * 
     * @type {string}
     * @memberof ScrapeSource
     */
    error?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ScrapeSource
     */
    startDate?: string;
    /**
     * 
     * @type {string}
     * @memberof ScrapeSource
     */
    endDate?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ScrapeSource
     */
    movieCount?: number;
    /**
     * 
     * @type {number}
     * @memberof ScrapeSource
     */
    torrentCount?: number;
}